common: settings: Add pydantic POC settings
